requirements.txt
requests
google-auth-oauthlib
google-api-python-client
google-auth
pyicloud-ipd
pillow
numpy
opencv-python
O365
python-dotenv
requests-oauthlib

